tests/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mongomock==4.1.2
mongomock_motor==0.0.28
motor==3.3.2
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.6.2
pytest==8.1.1
pytest-asyncio==0.23.5.post1
python-dotenv==1.0.1
sentinels==1.0.0
setuptools==68.2.2
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
wheel==0.41.2
